# qa-shop/api/openapi.yaml
openapi: 3.0.3
...
info:
  title: QA Shop Practice API
  version: "1.0.0"
  description: |
    Test otomasyonu pratiği için gerçek bir e-ticaret API'si. Mock değil:
    veri kalıcı, işlemler transaction içinde, stok gerçekten düşüyor.

    **Sandbox anahtarı göndermezsen** demo verisine salt okunur bağlanırsın —
    `GET` uçları çalışır, yazma denemeleri `401` döner. İlk denemeyi
    kolaylaştırmak için böyle.

    **Kendi alanını açmak için:** `POST /sandbox` → dönen `apiKey` değerini
    `X-Sandbox-Key` başlığında gönder. Alan 7 gün sonra otomatik silinir.

    **Her test koşumundan önce** `POST /sandbox/reset` çağır — temiz durum
    otomasyonun en temel disiplinidir ve burada bir uç olarak hazır.

    Demo hesabı: `demo@qashop.test` / `Password123!`
  contact:
    name: LearnQA.dev
    url: https://learnqa.dev
servers:
  - url: http://localhost:4000
    description: Lokal Docker (docker compose up)
...
security:
  - SandboxKey: []
paths:
...
  /api/v1/addresses:
    get:
      tags: [Adres]
      summary: 🔒 Adreslerim
      description: Varsayılan adres listenin başında döner.
      security:
        - SandboxKey: []
          BearerAuth: []
      responses:
        "200":
          description: Adres listesi
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  addresses:
                    type: array
                    items: { $ref: "#/components/schemas/Address" }
    post:
      tags: [Adres]
      summary: 🔒 Adres ekle
      description: |
        İLK adres otomatik olarak varsayılan olur — kullanıcı hiç seçim yapmadan
        checkout'a geldiğinde "varsayılan adres yok" durumu oluşmasın diye.

        `isDefault: true` gönderildiğinde diğer adreslerin varsayılanlığı DÜŞER.
        Yani bu istek, gövdede adı geçmeyen satırları da değiştirir; cevaba bakan
        bir test bunu göremez, veritabanına bakan bir test görür.
      security:
        - SandboxKey: []
          BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [line1, city]
              properties:
                label: { type: string, example: ev }
                line1: { type: string, minLength: 3 }
                city: { type: string, minLength: 2 }
                country: { type: string, pattern: "^[A-Za-z]{2}$", default: TR }
                postalCode: { type: string, nullable: true }
                isDefault: { type: boolean }
      responses:
        "201":
          description: Adres oluşturuldu
          content:
            application/json:
              schema:
                type: object
                properties:
                  address: { $ref: "#/components/schemas/Address" }
        "422":
          description: "`MISSING_FIELD`, `INVALID_FIELD` veya `INVALID_COUNTRY`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/addresses/{id}:
    patch:
      tags: [Adres]
      summary: 🔒 Adres güncelle
      description: Kısmi güncelleme — yalnızca gönderilen alanlar değişir.
      security:
        - SandboxKey: []
          BearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label: { type: string }
                line1: { type: string }
                city: { type: string }
                country: { type: string }
                postalCode: { type: string, nullable: true }
                isDefault: { type: boolean }
      responses:
        "200":
          description: Güncellenmiş adres
          content:
            application/json:
              schema:
                type: object
                properties:
                  address: { $ref: "#/components/schemas/Address" }
        "422":
          description: "`EMPTY_PATCH` — güncellenecek alan gönderilmedi"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403": { description: Adres sana ait değil }
        "404": { description: Adres bulunamadı }
    delete:
      tags: [Adres]
      summary: 🔒 Adres sil
      description: |
        Varsayılan adres silinirse kalanlardan biri varsayılan yapılır. Aksi
        hâlde kullanıcının hiç varsayılanı kalmaz ve bu ancak checkout'ta fark
        edilirdi.
      security:
        - SandboxKey: []
          BearerAuth: []
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Silindi }
        "403": { description: Adres sana ait değil }
        "404": { description: Adres bulunamadı }

  # ─────────────────────────────────── Yorum ─────────────────────────────────
...
components:
  securitySchemes:
    SandboxKey:
      type: apiKey
      in: header
      name: X-Sandbox-Key
      description: |
        Hangi veri alanına bağlanacağını belirler. `POST /api/v1/sandbox` ile
        alınır. Gönderilmezse demo verisine SALT OKUNUR bağlanılır.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `POST /api/v1/auth/login` ile alınır. Gerçek JWT formatındadır —
        jwt.io'ya yapıştırıp içindeki `sub`, `sandbox`, `jti`, `exp`
        alanlarını görebilirsin.
...
  schemas:
    Error:
      type: object
      description: |
        Hata gövdesi HER uçta aynı şekildedir. Her endpoint kendi hata şeklini
        döndürseydi test kodu her uç için ayrı yazılır ve kırılgan olurdu.
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: OUT_OF_STOCK }
            message: { type: string, example: Stokta yalnızca 2 adet var }
            details:
              type: object
              nullable: true
              example: { requested: 5, available: 2 }
        correlationId:
          type: string
          example: req-8f3c1e2a-...
          description: Cevabın `X-Correlation-Id` başlığında da döner.
...
    Address:
      type: object
      properties:
        id: { type: integer, example: 12 }
        label: { type: string, example: ev }
        line1: { type: string, example: "Ataturk Cad. No 12 D3" }
        city: { type: string, example: Istanbul }
        country: { type: string, example: TR }
        postal_code: { type: string, nullable: true, example: "34100" }
        is_default:
          type: boolean
          description: Kullanıcı başına yalnızca BİR adres varsayılan olabilir.
...